# backend/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: WebTui Chat Backend API
  version: 0.9.0
  description: Contract nền cho backend WebTui Chat. Tài liệu API được cập nhật theo từng phase backend.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /api/v1/workspaces/{workspace_id}:
    get:
      summary: Xem chi tiết workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "200":
          description: Chi tiết workspace
    patch:
      summary: Cập nhật workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "200":
          description: Workspace đã cập nhật
    delete:
      summary: Xóa mềm workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "204":
          description: Đã xóa mềm
...
  /api/v1/integrations/messages:
    post:
      security: []
      summary: Gửi message bằng API token
      description: "Dùng header `Authorization: Bearer {api_token}`. Token cần scope `message.write`."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntegrationMessageRequest"
      responses:
        "201":
          description: Message tích hợp đã được gửi
  /api/v1/workspaces/{workspace_id}/tickets:
    get:
      summary: List workspace tickets
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - name: status
          in: query
          schema:
            type: string
            enum: [open, pending, resolved, closed]
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Ticket list
          content:
            application/json:
              schema:
                type: object
                properties:
                  tickets:
                    type: array
                    items:
                      $ref: "#/components/schemas/Ticket"
    post:
      summary: Create workspace ticket
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketRequest"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /api/v1/workspaces/{workspace_id}/tickets/{ticket_id}:
    get:
      summary: Get workspace ticket
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/TicketID"
      responses:
        "200":
          description: Ticket detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
    patch:
      summary: Update workspace ticket lifecycle
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/TicketID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTicketRequest"
      responses:
        "200":
          description: Ticket updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
...
  /api/v1/workspaces/{workspace_id}/incoming-webhooks:
    get:
      summary: Danh sách incoming webhook
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "200":
          description: Danh sách incoming webhook
    post:
      summary: Tạo incoming webhook
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncomingWebhookRequest"
      responses:
        "201":
          description: Incoming webhook được tạo. Secret chỉ trả về một lần.
  /api/v1/hooks/incoming/{webhook_id}:
    post:
      security: []
      summary: Dispatch incoming webhook để gửi message vào channel
      parameters:
        - $ref: "#/components/parameters/WebhookID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomingWebhookDispatchRequest"
      responses:
        "201":
          description: Incoming webhook đã tạo message
  /api/v1/workspaces/{workspace_id}/outgoing-webhooks:
    get:
      summary: Danh sách outgoing webhook
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "200":
          description: Danh sách outgoing webhook
    post:
      summary: Tạo outgoing webhook
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOutgoingWebhookRequest"
      responses:
        "201":
          description: Outgoing webhook được tạo. Secret chỉ trả về một lần.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    WorkspaceID:
      name: workspace_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    WebhookID:
      name: webhook_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    TicketID:
      name: ticket_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    CreateIncomingWebhookRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        channel_id:
          type: string
          format: uuid
          description: Channel mặc định cho incoming webhook.
    IncomingWebhookDispatchRequest:
      type: object
      required: [body]
      properties:
        secret:
          type: string
          description: Có thể gửi qua body hoặc header X-WebTui-Webhook-Secret.
        channel_id:
          type: string
          format: uuid
          description: Ghi đè channel mặc định nếu webhook cho phép.
        body:
          type: string
        metadata:
          type: object
          additionalProperties: true
    CreateOutgoingWebhookRequest:
      type: object
      required: [name, target_url]
      properties:
        name:
          type: string
        target_url:
          type: string
          format: uri
        event_types:
          type: array
          items:
            type: string
          description: Rỗng nghĩa là nhận tất cả event.
    IntegrationMessageRequest:
      type: object
      required: [channel_id, body]
      properties:
        channel_id:
          type: string
          format: uuid
        body:
          type: string
        metadata:
          type: object
          additionalProperties: true
...
    Ticket:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        channel_id:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, pending, resolved, closed]
        priority:
          type: string
          enum: [low, normal, high, urgent]
        created_by:
          type: string
          format: uuid
          nullable: true
        assigned_to:
          type: string
          format: uuid
          nullable: true
        resolved_at:
          type: string
          format: date-time
          nullable: true
        closed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateTicketRequest:
      type: object
      required: [title]
      properties:
        channel_id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 160
        description:
          type: string
        priority:
          type: string
          enum: [low, normal, high, urgent]
        assigned_to:
          type: string
          format: uuid
    UpdateTicketRequest:
      type: object
      properties:
        channel_id:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
          maxLength: 160
        description:
          type: string
        status:
          type: string
          enum: [open, pending, resolved, closed]
        priority:
          type: string
          enum: [low, normal, high, urgent]
        assigned_to:
          type: string
          format: uuid
          nullable: true
...